fix: catch errors from mounted app
The mounted application will start a response if an exception is raised from within, and then re-raise.

If the re-raised exception is allowed to propagate, a new response will be started by the litestar handler, causing an error.

This PR catches any exception raised from within the mounted application and logs it.