feat: custom ModelView for AA Audit Model types.
Advanced-Alchemy (AA) provides variants of base models that include `created_at` and `updated_at` fields which enforce
that `tzinfo` is set on the `datetime` instance passed through to SQLAlchemy.

When a model is created via the SQLAdmin UI, the `created_at` and `updated_at` fields default to the current time in UTC,
however, the `tzinfo` property of the `datetime` is not set.

This PR defines a custom `ModelView` class that ensures the `tzinfo` property is set on `datetime`
instances when the form field represents an AA `DateTimeUTC` field.